Compose scraper which reports to Datadog
We have implemented a scraper for the metrics of our Compose cluster
(as used for production.) This uses the Scrapey library for Python to
perform the HTTPS requests, extract the relevant elements and extract
the RAM and Disk metrics for each host.

This is wrapped into an app to be deployed on the PaaS that performs
the scraping process periodically. Metrics are logged to Datadog.

The underlying HTML representation is non-semantic and seems likely to
change at a later date. We have broken down the scraping process into
`ClusterScraper` and `HostScraper` classes, which are designed to
resiliently extract as much information as possible even if one host
or one host's RAM information can't be extracted.

If information cannot be extracted, the scraper reports this to a
Datadog service check. Any missing data will be logged as a Warning.
If all data is missing (e.g., the page can no longer be scraped or no
hosts were extracted) this is logged as an Error.

A separate pull request on paas-cf integrates this into the PaaS:
https://github.com/alphagov/paas-cf/pull/1031