Check environment variables on startup
Previously the required environment variables were only fetched as
required by the subprocess. This gave poor error messages when they were
missing.

This changes it so that required variables are checked during execution
of the parent so we can bail with better errors early.

We also expose the CHECK_INTERVAL_SECONDS environment variable to make
it simpler to alter the time between scraping attempts