Run as a script using CrawlRunner rather than subprocess
Shelling out to a scrapey subprocess was giving poor errors and made
exiting the subprocess via SIGINT problimatic.

This changes to use the CrawlRunner helper to run the crawler in the
twisted reactor event loop manually.

As part of this change the global settings config was retired in favour
of building the settings dict within main to make a clear single
entrypoint for mapping of environment variables to settings.

The LOG_LEVEL setting was also exposed to environment and now defaults
to 'WARNING' to reduce the noise.